Create FeaturesView @Composable.
