+++ Demo/src/main/java/com/paypal/android/ui/features/FeaturesFragment.kt
@@ -39,35 +39,7 @@ import com.paypal.android.R
 
 class FeaturesFragment : Fragment() {
 
-    companion object {
-        // TODO: use a color scheme to keep UI consistent
-        val backgroundColor = Color(248, 249, 243)
-        val featureTitleColor = Color(20, 29, 47)
-        val chevronTint = Color(138, 137, 142)
-    }
-
-    enum class Feature(@StringRes val stringRes: Int) {
-        CARD_APPROVE_ORDER(R.string.feature_approve_order),
-        CARD_VAULT(R.string.feature_vault),
-        PAYPAL_WEB(R.string.feature_paypal_web),
-        PAYPAL_BUTTONS(R.string.feature_paypal_buttons),
-        PAYPAL_NATIVE(R.string.feature_paypal_native)
-    }
-
-    private val cardFeatures = listOf(
-        Feature.CARD_APPROVE_ORDER,
-        Feature.CARD_VAULT
-    )
-
-    private val payPalWebFeatures = listOf(
-        Feature.PAYPAL_WEB,
-        Feature.PAYPAL_BUTTONS
-    )
-
-    private val payPalNativeFeatures = listOf(
-        Feature.PAYPAL_NATIVE
-    )
-
+    @ExperimentalFoundationApi
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
@@ -79,12 +51,7 @@ class FeaturesFragment : Fragment() {
                 Surface(
                     modifier = Modifier.fillMaxSize()
                 ) {
-                    PaymentMethodsView(
-                        cardFeatures = cardFeatures,
-                        payPalWebFeatures = payPalWebFeatures,
-                        payPalNativeFeatures = payPalNativeFeatures,
-                        onFeatureSelected = ::onFeatureSelected,
-                    )
+                    FeaturesView(onFeatureSelected = ::onFeatureSelected)
                 }
             }
         }
@@ -115,124 +82,15 @@ class FeaturesFragment : Fragment() {
         findNavController().navigate(action)
     }
 
-    @Composable
-    @OptIn(ExperimentalFoundationApi::class)
-    fun PaymentMethodsView(
-        cardFeatures: List<Feature>,
-        payPalWebFeatures: List<Feature>,
-        payPalNativeFeatures: List<Feature>,
-        onFeatureSelected: (Feature) -> Unit,
-    ) {
-        LazyColumn(
-            modifier = Modifier
-                .background(backgroundColor)
-                .padding(horizontal = 16.dp)
-        ) {
-            stickyHeader {
-                FeatureGroupHeader(text = "Card")
-            }
-            item {
-                FeatureOptions(cardFeatures, onFeatureSelected = onFeatureSelected)
-            }
-            stickyHeader {
-                FeatureGroupHeader("PayPal Web")
-            }
-            item {
-                FeatureOptions(payPalWebFeatures, onFeatureSelected = onFeatureSelected)
-            }
-            stickyHeader {
-                FeatureGroupHeader("PayPal Native")
-            }
-            item {
-                FeatureOptions(payPalNativeFeatures, onFeatureSelected = onFeatureSelected)
-            }
-        }
-    }
 
+    @ExperimentalFoundationApi
     @Preview
     @Composable
     fun PaymentMethodsViewPreview() {
         MaterialTheme {
             Surface(modifier = Modifier.fillMaxSize()) {
-                PaymentMethodsView(
-                    cardFeatures = cardFeatures,
-                    payPalWebFeatures = payPalWebFeatures,
-                    payPalNativeFeatures = payPalNativeFeatures,
-                    onFeatureSelected = {},
-                )
+                FeaturesView(onFeatureSelected = {})
             }
         }
-    }
-
-    @Composable
-    fun FeatureOptions(
-        features: List<Feature>,
-        onFeatureSelected: (Feature) -> Unit,
-    ) {
-        Card(
-            shape = CardDefaults.elevatedShape,
-            elevation = CardDefaults.elevatedCardElevation()
-        ) {
-            features.forEachIndexed { index, feature ->
-                FeatureView(
-                    feature = feature,
-                    isLast = (index == features.lastIndex),
-                    onClick = { onFeatureSelected(feature) }
-                )
-            }
-        }
-    }
-
-    @Composable
-    fun FeatureView(
-        feature: Feature,
-        isLast: Boolean,
-        onClick: () -> Unit
-    ) {
-        val chevronPainter = painterResource(id = R.drawable.chevron)
-        Column(
-            modifier = Modifier
-                .background(Color.White)
-                .fillMaxWidth()
-                .clickable(onClick = onClick)
-        ) {
-            Row {
-                Text(
-                    text = stringResource(id = feature.stringRes),
-                    color = Color.Black,
-                    style = MaterialTheme.typography.titleLarge,
-                    modifier = Modifier
-                        .weight(1.0f)
-                        .padding(vertical = 16.dp, horizontal = 20.dp)
-                )
-                Icon(
-                    painter = chevronPainter,
-                    contentDescription = null,
-                    tint = chevronTint,
-                    modifier = Modifier
-                        .size(14.dp)
-                        .align(Alignment.CenterVertically)
-                )
-                Spacer(modifier = Modifier.size(20.dp))
-            }
-            if (!isLast) {
-                Divider(
-                    color = backgroundColor,
-                    modifier = Modifier.padding(start = 16.dp)
-                )
-            }
-        }
-    }
-
-    @Composable
-    fun FeatureGroupHeader(text: String) {
-        Spacer(modifier = Modifier.size(24.dp))
-        Text(
-            text = text,
-            color = featureTitleColor,
-            fontWeight = FontWeight.Bold,
-            style = MaterialTheme.typography.titleLarge,
-        )
-        Spacer(modifier = Modifier.size(12.dp))
     }
 }
